requirements.txt
django==5.2.8
python-environ==0.4.54
django-cors-headers==4.7.0
psycopg2-binary==2.9.11
google-generativeai==0.8.3
celery==5.3.4
django-celery-beat==2.8.1
django-celery-results==2.6.0
redis==5.0.1
gunicorn
requests==2.32.3
pillow==12.0.0
beautifulsoup4==4.12.3
ipython
djangorestframework
djangorestframework-simplejwt==5.3.0